Changed a method that was left behind in removing an employee and asking a redundant question.
Changed the delete employee to refine the search a little better. Changed the employee database id to first name and employee number. Admin may not know the unique key from the database.

Changed the information that gets passed back on reading an employee. Added descriptive words to each attribute of an employee that gets passed back so it looks clean.